hvue.Method(); currency-input example
- Added hvue.Method() to add a naked closure as a method, instead of using
  MethodsOf, which forces you to define a type and then methods on it.
- Added the currency-input example

I'm not sure I like the idea of hvue.Method().  One of the benefits of
MethodsOf is that methods are named and known to the compiler (and other tools),
instead of just closures pointed to by strings.

But on the other hand I guess I shouldn't make it *impossible* to use regular
closures.  So keep it for now.