First quick-and-dirty pass at a wasm port
- Used "github.com/gopherjs/gopherwasm/js" to try to make it easier to
  use with both go/wasm and GopherJS.
- Only 01-introduction tested, and only with wasm, not GopherJS
- You have to use the included server to run with wasm, because
  "gopherjs serve" does not set the right mime type for .wasm files.
  The server is fairly stupid because I was lazy; in particular, it only
  serves wasm files correctly (with the right mime type) if they're in
  examples/wasm.

Still some bugs, still kludgy, but most of examples/01-introduction
work.  Actually they *all* work, they just cause Vue to throw warnings,
since the data functions aren't working correctly.

Speaking of which: Vue.js expects to be able to call functions
synchronously when creating components (among other places).  There is
currently *no way* to call a go/wasm function synchronously from
JavaScript.  So that's going to be broken until synchronous function
support comes in.

I *did* find a way to write callbacks in such a way so as to capture the
"this" value of the callback and pass that on to Go, so I'm happy about
that.